Rearranging some classes to its own files
