+++ source/custom_controls/control_strategy_TE.py
@@ -1,6 +1,8 @@
 
 from math import floor, ceil, fmod
 
+from dynamic_price_control.charge_controller import charge_controller
+
 from Caldera_globals import L2_control_strategies_enum, SE_setpoint, timeseries, active_CE
 from global_aux import Caldera_message_types, OpenDSS_message_types, input_datasets, container_class
 from control_templates import typeA_control
@@ -9,7 +11,6 @@ from Caldera_ICM_Aux import CP_interface_v2
 import os 
 import pandas as pd
 import numpy as np
-from typing import List
 import matplotlib.pyplot as plt
 import json
 
@@ -101,14 +102,13 @@ class control_strategy_TE(typeA_control):
         forecasted_cost = self.controller.get_forecasted_cost_at_time_sec(next_control_starttime_sec)
         actual_cost = self.controller.get_actual_cost_at_time_sec(next_control_starttime_sec)
         
-        print("Control Strategy forecasted_cost : ", forecasted_cost)
-        print("Control Strategy actual_cost : ", actual_cost)
+        #print("Control Strategy forecasted_cost : ", forecasted_cost)
+        #print("Control Strategy actual_cost : ", actual_cost)
         
         tolerance = 0.10        # 10 percent
         cost_deviated_from_forecast = ((actual_cost - forecasted_cost) / forecasted_cost > tolerance)        
         print("Control Strategy cost_deviated_from_forecast : ", cost_deviated_from_forecast)
         
-        print(Caldera_state_info_dict[Caldera_message_types.get_active_charge_events_by_extCS].keys())
         CEs_all = Caldera_state_info_dict[Caldera_message_types.get_active_charge_events_by_extCS][self.cs_id]
 
         active_SEs = []
@@ -158,1173 +158,3 @@ class control_strategy_TE(typeA_control):
         # DSS_control_info_dict must be a dictionary with OpenDSS_message_types as keys.
         # If either value has nothing to return, return an empty dictionary.
         return (Caldera_control_info_dict, DSS_control_info_dict)
-
-class charge_controller:
-    '''
-    Description:
-        charge_controller is the controller that decides cheapest time for charge to occur based on cost of electricity.
-    '''
-    
-    def __init__(self, io_dir, starttime_sec : float, endtime_sec : float, timestep_sec : float, SE_ids : List[int]):
-        '''
-        Description:
-            constuctor initializes the charge controller, allocates the controller_2Darr that maintains the status of charge events being controlled.
-        '''
-        
-        self.plot = True
-        self.debug_plot = False
-        self.plots = set()
-        self.input_folder = io_dir.inputs_dir
-        self.figures_folder = io_dir.figures_dir
-        
-        self.forecast_duration_sec = 12*3600            # Max forecast of 12 hours
-        self.controller_starttime_sec = starttime_sec   
-        self.controller_endtime_sec = endtime_sec + self.forecast_duration_sec
-        self.controller_timestep_sec = timestep_sec
-        self.charge_profile_timestep_sec = 60           # 1 minute timestep
-        
-        self.use_cost_forecaster_v3 = True
-        
-        # CP_interface_v2 generates charge profiles
-        self.charge_profiles = CP_interface_v2(self.input_folder)
-        
-        if self.use_cost_forecaster_v3:
-            self.cost_forecaster = TE_cost_forecaster_v3(os.path.join(self.input_folder, "TE_inputs"), self.figures_folder, self.plot)
-        else:
-            
-            forecast_file = os.path.join(self.input_folder, "TE_inputs", "forecast.csv")
-            actual_file = os.path.join(self.input_folder, "TE_inputs", "actual.csv")
-            cost_file = os.path.join(self.input_folder, "TE_inputs", "generation_cost.json")
-        
-            # cost_forcaster contains the cost of energy
-            self.cost_forecaster = TE_cost_forecaster_v2(forecast_file, actual_file, cost_file, self.figures_folder, self.plot)
-        
-        # controller_2Darr keeps track of what SE's should be controlled at any given time
-        num_SEs = len(SE_ids)
-        num_steps = ceil((self.controller_endtime_sec - self.controller_starttime_sec)/ self.controller_timestep_sec) + 1
-        self.controller_2Darr = np.full((num_SEs, num_steps), False)
-        
-        # mappings of SE_id into controller_2Darr
-        self.SE_id_to_controller_index_map = {}
-        self.controller_index_to_SE_id_map = {}
-        
-        for (idx, SE_id) in enumerate(SE_ids):
-            self.SE_id_to_controller_index_map[SE_id] = idx
-            self.controller_index_to_SE_id_map[idx] = SE_id
-                
-    #def log_controller(self):
-    #    df = pd.DataFrame()
-    #    time_arr = np.array([i for i in range(self.controller_starttime_sec, self.controller_endtime_sec + self.controller_timestep_sec, self.controller_timestep_sec)])
-    #    df["time"] = time_arr/3600.0
-    #    df["SE_100239126"] = self.controller_2Darr[self.SE_id_to_controller_index_map[100239126], :].astype(int)
-    #    df.to_csv(os.path.join(self.input_folder, "controller_state.csv"), index = False)
-        
-    def get_forecasted_cost_at_time_sec(self, time_sec : float) -> float:
-        '''
-        Description:
-            Given a time return the forecasted cost at that time from cost forecaster.
-        '''
-        
-        return self.cost_forecaster.get_forecasted_cost_at_time_sec(time_sec)
-    
-    def get_actual_cost_at_time_sec(self, time_sec : float) -> float:
-        '''
-        Description:
-            Given a time return the actual cost at that time from cost forecaster.
-        '''
-
-        return self.cost_forecaster.get_actual_cost_at_time_sec(time_sec)
-    
-    
-    def get_time_idx_from_time_sec(self, time_sec : float) -> int:
-        '''
-        Description:
-            Given a time return the time index in controller_2Darr.
-        '''
-        
-        return int((time_sec - self.controller_starttime_sec)/self.controller_timestep_sec)
-        
-        
-    def recalculate_active_charge_event(self, next_control_starttime_sec : float, active_charge_event : active_CE):
-        '''
-        Description:
-            Resets the previously computed control actions for the SE and computes control actions again based on new available cost information.
-        '''
-        
-        SE_id = active_charge_event.SE_id
-        
-        # clear old optimized charge event profile
-        idx = self.SE_id_to_controller_index_map[SE_id]
-        start_time_idx = self.get_time_idx_from_time_sec(next_control_starttime_sec)
-        self.controller_2Darr[idx, start_time_idx:] = 0
-        
-        # Compute for this charge event again
-        self.add_active_charge_event(next_control_starttime_sec, active_charge_event)
-    
-    def add_active_charge_event(self, next_control_starttime_sec : float, active_charge_event : active_CE) -> None:
-        '''
-        Description:
-            Adds charge events control actions to the controller by updating controller_2Darr with cheapest 
-            times to charge
-        '''
-        
-        # get the required info from active_charge_event
-        SE_id = active_charge_event.SE_id
-        CE_id = active_charge_event.charge_event_id
-        supply_equipment_type = active_charge_event.supply_equipment_type
-        vehicle_type = active_charge_event.vehicle_type
-        departure_unix_time = active_charge_event.departure_unix_time
-        arrival_SOC = active_charge_event.arrival_SOC
-        departure_SOC = active_charge_event.departure_SOC
-        now_soc = active_charge_event.now_soc
-        charge_time_remaining_hrs = active_charge_event.min_remaining_charge_time_hrs
-        charge_time_total_hrs = active_charge_event.min_time_to_complete_entire_charge_hrs
-        
-        #-------------------------------------------
-        #       Build time parameters
-        #-------------------------------------------
-        start_time_sec = next_control_starttime_sec
-        end_time_sec = min(next_control_starttime_sec + self.forecast_duration_sec, departure_unix_time - fmod(departure_unix_time, self.controller_timestep_sec))
-        
-        if(end_time_sec <= start_time_sec):
-            print("Charge event {} too small. Cannot be controlled".format(CE_id))
-            return
-        
-        # round up to the nearest self.controller_timestep_sec
-        charge_time_remaining_in_controller_step_sec = ceil(charge_time_remaining_hrs * 3600 / self.controller_timestep_sec) * self.controller_timestep_sec
-        
-        # Number of steps controller needs to charge the vehicle
-        num_steps_to_charge_by_controller = int(charge_time_remaining_in_controller_step_sec / self.controller_timestep_sec)
-
-        #-------------------------------------------
-        #       Build cost profile timeseries
-        #-------------------------------------------
-        
-        if self.use_cost_forecaster_v3:
-            cost_profile = self.cost_forecaster.get_cost_for_time_range(next_control_starttime_sec, start_time_sec, end_time_sec, self.controller_timestep_sec)
-        else:
-            cost_profile = self.cost_forecaster.get_cost_for_time_range(start_time_sec, end_time_sec, self.controller_timestep_sec)
-        
-        #-------------------------------------------
-        #       Select least cost times 
-        #-------------------------------------------
-        
-        cost_profile_indices_sorted = sorted(range(len(cost_profile.data)), key = lambda k: cost_profile.data[k])
-        
-        # select first n cheapest cost profiles 
-        # TODO : make the cost profile as contiguous as possible
-        cost_profile_indeces_cheapest = cost_profile_indices_sorted[:num_steps_to_charge_by_controller]
-        
-        #-------------------------------------------
-        #       Update controller_2Darr 
-        #-------------------------------------------
-        
-        for cost_profile_index in cost_profile_indeces_cheapest:
-            profile_time_sec = cost_profile.get_time_from_index_sec(cost_profile_index)
-            control_time_index = self.get_time_idx_from_time_sec(profile_time_sec)          
-            self.controller_2Darr[self.SE_id_to_controller_index_map[SE_id], control_time_index] = True
-        
-        if self.debug_plot == True:
-            
-            #-------------------------------------------
-            #       Build Charge profile timeseries
-            #-------------------------------------------
-        
-            # create charge_profile
-            all_charge_profile_data = self.charge_profiles.create_charge_profile_from_model(self.charge_profile_timestep_sec, vehicle_type, supply_equipment_type, now_soc, departure_SOC, 1000, {}, {})
-        
-            # create 15 min buckets, with energy consumed in each bucket.
-            num_bins_to_aggregate = int(self.controller_timestep_sec / self.charge_profile_timestep_sec)
-        
-            charge_profile_data = (np.add.reduceat(all_charge_profile_data.P3_kW, np.arange(0, len(all_charge_profile_data.P3_kW), num_bins_to_aggregate))/num_bins_to_aggregate)*(self.controller_timestep_sec/3600.0)
-        
-            # convert charge_profile_data to timeseries
-            charge_profile = timeseries(start_time_sec, self.controller_timestep_sec, charge_profile_data)
-            
-            #-------------------------------------------
-            
-            start_idx = self.get_time_idx_from_time_sec(start_time_sec)  
-            end_idx = self.get_time_idx_from_time_sec(end_time_sec)            
-            profile_size = end_idx - start_idx
-            
-            time_profile = np.arange(start_time_sec, end_time_sec, self.controller_timestep_sec)/3600.0
-            charge_profile = charge_profile.data[:profile_size] + [0.0]*(profile_size - len(charge_profile.data))     # make charge profile length same as other profiles by adding 0's to the end
-            cost_profile = cost_profile.data
-            control_profile = self.controller_2Darr[self.SE_id_to_controller_index_map[SE_id], start_idx:end_idx]
-               
-            fig, axes = plt.subplots(3, 1, figsize=(25, 15))
-            
-            ax = axes[0]
-            ax.plot(time_profile, charge_profile)
-            ax.set_xlabel("Time (hrs)")
-            ax.set_ylabel("Power (kW)")
-            ax.set_title("Charge profile")
-            ax.set_xlim(time_profile[0], time_profile[-1] + self.controller_timestep_sec/3600.0)            
-            ax.set_xticks(time_profile)
-            ax.grid(True, which='both', axis='both')
-
-            ax = axes[1]
-            ax.step(time_profile, cost_profile, where = 'post' )
-            ax.set_xlabel("Time (hrs)")
-            ax.set_ylabel("Cost ($$$)")
-            ax.set_title("Forecasted cost profile")
-            ax.set_xlim(time_profile[0], time_profile[-1] + self.controller_timestep_sec/3600.0)
-            ax.set_xticks(time_profile)
-            ax.grid(True, which='both', axis='both')
-
-            
-            ax = axes[2]
-            cmap = plt.get_cmap('viridis')
-            ax.bar(time_profile, align='edge', height=1, width = time_profile[1] - time_profile[0], color=cmap(np.array(control_profile, dtype = float)))
-            ax.set_xlabel("Time (hrs)")
-            ax.set_title("Forecasted control profile")
-            ax.tick_params(left = False)
-            ax.set_xlim(time_profile[0], time_profile[-1] + self.controller_timestep_sec/3600.0)
-            ax.set_ylim(0, 1)
-            ax.set_xticks(time_profile)
-            ax.set_yticks([])
-            ax.grid()
-            
-            plt.subplots_adjust(hspace=0.4)
-            base_name = 'CE_{}_plot_{:.2f}_'.format(CE_id, next_control_starttime_sec/3600.0)
-            
-            suffix = 0
-            while "{}{}".format(base_name, str(suffix)) in self.plots:
-                suffix += 1
-            
-            plot_name = "{}{}".format(base_name, str(suffix))
-            self.plots.add(plot_name)
-            
-            os.makedirs(self.figures_folder, exist_ok=True)
-            plt.savefig(os.path.join(self.figures_folder, plot_name+".png"))
-   
-            #self.plot = False
-               
-    def get_SE_setpoints(self, next_control_timestep_sec : float, active_SEs : List[int] ) -> List[SE_setpoint]:
-        '''
-        Description:
-            Looks up controller_2Darr to see which SEs needs to charge at the specific time
-        '''
-        
-        time_index = floor((next_control_timestep_sec - self.controller_starttime_sec)/ self.controller_timestep_sec)
-        
-        # check all rows with active charge events that need to charge
-        SE_indexes_to_charge = np.array(np.where(self.controller_2Darr[:, time_index] == True))
-        
-        PQ_setpoints = []
-        for SE_id in active_SEs:
-            X = SE_setpoint()
-            X.SE_id = SE_id
-            
-            if np.any(SE_indexes_to_charge == self.SE_id_to_controller_index_map[SE_id]):
-                X.PkW = 1000
-            else:
-                X.PkW = 0
-
-            PQ_setpoints.append(X)
-        
-        return PQ_setpoints
-    
-class TE_cost_forecaster_v2():
-    '''
-    Description:
-        The TE_cost_forecaster will read the input file and compute costs and loads forecasted cost and actual cost as timeseries data 
-    '''
-    
-    def __init__(self, forecast_input_file : str, actual_input_file : str, cost_input_file : str, figures_folder : str, plot : bool) -> None:
-        '''
-        Description:
-            Loads data, computes cost and stores as timeseries data
-        '''
-        
-        self.figures_folder = figures_folder
-        self.plot = plot
-        
-        # Load forecast and actual generation as dfs
-        df_forecast = pd.read_csv(forecast_input_file)
-        df_actual = pd.read_csv(actual_input_file)
-        
-        # Load costs as json dictionaray 
-        with open(cost_input_file, "r") as f_cost:
-            cost_json = json.load(f_cost)
-
-        # Three solver methods available linear, steep_cubic and inverse_s
-        solver_method = cost_json["cost_function"]
-        self.forecasted_cost_profile = self.solver(solver_method, df_forecast, cost_json)
-        self.actual_cost_profile = self.solver(solver_method, df_actual, cost_json)
-        
-        ## Error Checks
-        if not len(self.forecasted_cost_profile.data) == len(self.actual_cost_profile.data):
-           raise ValueError('ERROR : Input forecasted_cost_profile data and actual_cost_profile should have same length')
-        
-        if not ((self.forecasted_cost_profile.data_timestep_sec * len(self.forecasted_cost_profile.data)) % 24*3600 == 0.0):
-           raise ValueError('ERROR : Input forecasted_cost_profile data should be multiple of 24 hours of data')
-        
-        if not ((self.forecasted_cost_profile.data_timestep_sec * len(self.actual_cost_profile.data)) % 24*3600 == 0.0):
-           raise ValueError('ERROR : Input actual_cost_profile data should be multiple of 24 hours of data')
-        
-        self.cost_profile_timestep_sec = self.forecasted_cost_profile.data_timestep_sec
-        self.cost_profile_length_sec = self.cost_profile_timestep_sec * len(self.forecasted_cost_profile.data)
-
-    def solver(self, solver_method, df, cost_data) -> timeseries:
-        '''
-        Description:
-            Applies cost functions to the generation data
-        '''
-        df.columns = [column.split("|")[0].strip() for column in df.columns.to_series()]
-        
-        df_type = None 
-        if df.columns[1] == "forecasted_demand":
-            df_type = "forecast"         
-        elif df.columns[1] == "actual_demand":
-            df_type = "actual"
-        else:
-            raise ValueError('ERROR : Second column in input csv file should be forecasted_demand or actual_demand')
-            
-        start_time_sec = round(df["time"][0] * 3600.0)
-        timestep_sec = round((df["time"][1] - df["time"][0])*3600)
-        
-        # Ignore first 2 columns time and forecasted_demand/actual_demand
-        gen_types = df.columns[2:].to_series()
-        
-        cost_usd_per_kWh = np.zeros(df.shape[0], dtype=float)
-        total_cost_usd = np.zeros(df.shape[0], dtype=float)
-        
-        individual_costs = {}
-        for gen_type in gen_types:
-            
-            gen_min = cost_data[gen_type]["gen_min"]            #MW
-            gen_max = cost_data[gen_type]["gen_max"]            #MW
-            cost_min = cost_data[gen_type]["cost_min"]          # USD per MWh
-            cost_max = cost_data[gen_type]["cost_max"]          # USD per MWh
-            
-            # No cost variation
-            if abs(cost_min - cost_max) < 0.001:      # cost_min == cost_max        # nuclear and solar come under this scenario
-                
-                individual_cost_function = np.full_like(total_cost_usd, cost_min, dtype=float)
-                total_cost_usd += individual_cost_function*df[gen_type]*timestep_sec/3600.0
-            
-            # Cost variation
-            else:                                                                   # Thermal scenario
-            
-                if solver_method == "linear":
-                    
-                    individual_cost_function = (df[gen_type] - gen_min)/(gen_max-gen_min)*(cost_max-cost_min)+ cost_min
-                    individual_costs[gen_type] = individual_cost_function
-                    
-                    total_cost_usd += individual_cost_function*df[gen_type]*timestep_sec/3600.0
-                
-                elif solver_method == "steep_cubic":
-                    
-                    dxdy_at_gen_min = 0
-                    dxdy_at_gen_max = 0.35
-
-                    
-                    A = [[gen_min**0, gen_min**1,   gen_min**2,     gen_min**3], 
-                         [0,          1*gen_min**0, 2*gen_min**1,   3*gen_min**2], 
-                         [gen_max**0, gen_max**1,   gen_max**2,     gen_max**3],
-                         [0,          1*gen_max**0, 2*gen_max**1,   3*gen_max**2]]
-
-                    b = [[cost_min],
-                         [dxdy_at_gen_min],
-                         [cost_max],
-                         [dxdy_at_gen_max]]
-                    
-                    C = solve(A, b)
-                    
-                    individual_cost_function = C[0][0] * df[gen_type]**0 + C[1][0] * df[gen_type]**1 + C[2][0] * df[gen_type]**2 + C[3][0] * df[gen_type]**3
-                    individual_costs[gen_type] = individual_cost_function
-                    
-                    total_cost_usd += individual_cost_function*df[gen_type]*timestep_sec/3600.0
-                    
-                elif solver_method == "inverse_s":
-                    
-                    dxdy_at_gen_min = 0.3
-                    dxdy_at_gen_max = 0.3
-
-
-                    A = [[gen_min**0, gen_min**1,   gen_min**2,     gen_min**3], 
-                         [0,          1*gen_min**0, 2*gen_min**1,   3*gen_min**2], 
-                         [gen_max**0, gen_max**1,   gen_max**2,     gen_max**3],
-                         [0,          1*gen_max**0, 2*gen_max**1,   3*gen_max**2]]
-
-                    b = [[cost_min],
-                         [dxdy_at_gen_min],
-                         [cost_max],
-                         [dxdy_at_gen_max]]
- 
-                    C = solve(A, b)
-
-                    individual_cost_function = C[0][0] * df[gen_type]**0 + C[1][0] * df[gen_type]**1 + C[2][0] * df[gen_type]**2 + C[3][0] * df[gen_type]**3
-                    individual_costs[gen_type] = individual_cost_function
-                    
-                    total_cost_usd += individual_cost_function*df[gen_type]*timestep_sec/3600.0
-                    
-                else:
-                    
-                    raise ValueError('ERROR : solver_method should be linear or steep_cubic or inverse_s')
-            
-            #if self.plot:
-            #    fig, ax = plt.subplots(1, 1, figsize=(25, 15))
-            #
-            #    ax.plot(df[gen_type], individual_cost_function)
-            #    ax.set_xlabel("Power | MW")
-            #    ax.set_ylabel("Cost | $ per MWh")
-            #
-            #    fig.savefig("fig_" + gen_type + ".png", dpi = 300)
-            
-        
-        total_MWh_per_timestep = df[gen_types].sum(axis=1) * timestep_sec/3600.0
-        cost_usd_per_MWh = total_cost_usd/total_MWh_per_timestep
-        cost_usd_per_kWh = cost_usd_per_MWh/1000.0
-        
-#        if self.plot:
-#            
-#            os.makedirs(self.figures_folder, exist_ok= True)
-#        
-#            start_idx = (1*24*3600) / timestep_sec
-#            end_idx = (2*24*3600) / timestep_sec
-#            
-#            fig, ax = plt.subplots(1, 1)
-#            ax.plot(df["time"], individual_costs["fossil_fuel"])
-#            ax.set_xlabel("Time | hrs")
-#            ax.set_ylabel("Cost | $ per kWh")
-#            ax.set_title("Time vs Cost for fossil fuel with {} pricing".format(solver_method))
-#            ax.grid()
-#            fig.savefig(os.path.join(self.figures_folder, "ff_" + df_type + "_" + solver_method + "_time_vs_cost.png"), dpi = 300)
-#        
-#            fig, ax = plt.subplots(1, 1)
-#            ax.scatter(df["fossil_fuel"], individual_costs["fossil_fuel"])
-#            ax.set_xlabel("Power | kW")
-#            ax.set_ylabel("Cost | $ per kWh")
-#            ax.grid()
-#            fig.savefig(os.path.join(self.figures_folder, "ff_" + df_type + "_" + solver_method + "_power_vs_cost.png"), dpi = 300)
-                
-        return timeseries(start_time_sec, timestep_sec, cost_usd_per_kWh)
-            
-    def error_check(self, starttime_sec : float, endtime_sec : float, req_timestep_sec : float) -> timeseries:
-        '''
-        Description:
-            Check if the parameters match up.
-        '''
-        if not ((self.cost_profile_timestep_sec == req_timestep_sec) or \
-               (self.cost_profile_timestep_sec % req_timestep_sec == 0.0) or \
-               (req_timestep_sec % self.cost_profile_timestep_sec == 0.0)):
-            
-            print("One of the three checks below needs to be True")
-            print("self.cost_profile_timestep_sec == req_timestep_sec:", self.cost_profile_timestep_sec == req_timestep_sec)
-            print("self.cost_profile_timestep_sec % req_timestep_sec == 0.0:", self.cost_profile_timestep_sec % req_timestep_sec == 0.0)
-            print("req_timestep_sec % self.cost_profile_timestep_sec == 0.0:", req_timestep_sec % self.cost_profile_timestep_sec == 0.0)
-            
-            raise ValueError('ERROR : timestep parameters to get_cost_for_time_range are incompatible')
-        
-        if (starttime_sec % self.cost_profile_timestep_sec != 0.0) or \
-           (endtime_sec % self.cost_profile_timestep_sec != 0.0) or \
-           (starttime_sec >= endtime_sec ) or \
-           (req_timestep_sec < 0):
-
-           print("starttime_sec % self.cost_profile_timestep_sec should be equal to 0 :", starttime_sec % self.cost_profile_timestep_sec)
-           print("endtime_sec %self.cost_profile_timestep_sec should be equal to 0: ", endtime_sec % self.cost_profile_timestep_sec)
-           print("starttime_sec >= endtime_sec should be false: ", starttime_sec >= endtime_sec)
-           print("req_timestep_sec < 0  should be false: ", req_timestep_sec < 0)
-            
-           raise ValueError('ERROR : parameters to get_cost_for_time_range are incompatible')
-
-        
-    def get_cost_for_time_range(self, starttime_sec : float, endtime_sec : float, req_timestep_sec : float) -> timeseries:
-        '''
-        Description:
-            Given a timerange and timestep, looksup the data and returns the cost. 
-            the timerange should match up with data timeperiod.
-        '''
-        
-        self.error_check(starttime_sec, endtime_sec, req_timestep_sec)
-
-        data = []
-        if (self.forecasted_cost_profile.data_timestep_sec >= req_timestep_sec):
-            for time_sec in range(int(starttime_sec), int(endtime_sec), int(req_timestep_sec)):
-            
-                # The assumption here is that for current time. in this case the starttime_sec, real time actual cost is available.
-                # beyond the current time, forecasted cost is used.
-                if time_sec == int(starttime_sec):
-                    cost = self.actual_cost_profile.get_val_from_time(time_sec % self.cost_profile_length_sec)
-                else:
-                    cost = self.forecasted_cost_profile.get_val_from_time(time_sec % self.cost_profile_length_sec)
-                
-                data.append(cost)
-                
-        elif (self.forecasted_cost_profile.data_timestep_sec < req_timestep_sec):
-            for time_sec in range(int(starttime_sec), int(endtime_sec), int(req_timestep_sec)):
-                
-                start_sec = time_sec
-                end_sec = time_sec + req_timestep_sec
-                divisor = req_timestep_sec/self.forecasted_cost_profile.data_timestep_sec
-                
-                total_cost = 0    
-                if(time_sec == int(starttime_sec)):
-                    # Use actual price in this section
-                    for subtime_sec in range(int(start_sec), int(end_sec), int(self.forecasted_cost_profile.data_timestep_sec)):
-                        total_cost += self.actual_cost_profile.get_val_from_time(subtime_sec % self.cost_profile_length_sec)
-                else:
-                    # Use forecasted price in this section
-                    for subtime_sec in range(int(start_sec), int(end_sec), int(self.forecasted_cost_profile.data_timestep_sec)):
-                        total_cost += self.forecasted_cost_profile.get_val_from_time(subtime_sec % self.cost_profile_length_sec)
-                
-                avg_cost = total_cost/divisor
-                data.append(avg_cost)
-                
-        else:
-            raise ValueError('ERROR : get_cost_for_time_range')
-
-        return timeseries(starttime_sec, req_timestep_sec, data)
-
-    def get_forecasted_cost_for_time_range(self, starttime_sec : float, endtime_sec : float, req_timestep_sec : float) -> timeseries:
-        '''
-        Description:
-            Given a timerange and timestep, looksup the data and returns the forecasted cost. 
-            the timerange should match up with data timeperiod.
-        '''
-        
-        self.error_check(starttime_sec, endtime_sec, req_timestep_sec)
-        
-        data = []
-        if (self.forecasted_cost_profile.data_timestep_sec >= req_timestep_sec):
-            for time_sec in range(int(starttime_sec), int(endtime_sec), int(req_timestep_sec)):
-            
-                cost = self.forecasted_cost_profile.get_val_from_time(time_sec % self.cost_profile_length_sec)
-                data.append(cost)
-                
-        elif (self.forecasted_cost_profile.data_timestep_sec < req_timestep_sec):
-            for time_sec in range(int(starttime_sec), int(endtime_sec), int(req_timestep_sec)):
-                
-                start_sec = time_sec
-                end_sec = time_sec + req_timestep_sec
-                divisor = req_timestep_sec/self.forecasted_cost_profile.data_timestep_sec
-                
-                total_cost = 0    
-                
-                for subtime_sec in range(int(start_sec), int(end_sec), int(self.forecasted_cost_profile.data_timestep_sec)):
-                    total_cost += self.forecasted_cost_profile.get_val_from_time(subtime_sec % self.cost_profile_length_sec)
-                
-                avg_cost = total_cost/divisor
-                data.append(avg_cost)
-                
-        else:
-            raise ValueError('ERROR : get_cost_for_time_range')
-
-        return timeseries(starttime_sec, req_timestep_sec, data)
-
-    def get_actual_cost_for_time_range(self, starttime_sec : float, endtime_sec : float, req_timestep_sec : float) -> timeseries:
-        '''
-        Description:
-            Given a timerange and timestep, looksup the data and returns the actual cost. 
-            the timerange should match up with data timeperiod.
-        '''
-        
-        self.error_check(starttime_sec, endtime_sec, req_timestep_sec)
-        
-        data = []
-        if (self.actual_cost_profile.data_timestep_sec >= req_timestep_sec):
-            for time_sec in range(int(starttime_sec), int(endtime_sec), int(req_timestep_sec)):
-            
-                cost = self.actual_cost_profile.get_val_from_time(time_sec % self.cost_profile_length_sec)    
-                data.append(cost)
-                
-        elif (self.actual_cost_profile.data_timestep_sec < req_timestep_sec):
-            for time_sec in range(int(starttime_sec), int(endtime_sec), int(req_timestep_sec)):
-                
-                start_sec = time_sec
-                end_sec = time_sec + req_timestep_sec
-                divisor = req_timestep_sec/self.forecasted_cost_profile.data_timestep_sec
-                
-                total_cost = 0
-                for subtime_sec in range(int(start_sec), int(end_sec), int(self.forecasted_cost_profile.data_timestep_sec)):
-                    total_cost += self.actual_cost_profile.get_val_from_time(subtime_sec % self.cost_profile_length_sec)
-                
-                avg_cost = total_cost/divisor
-                data.append(avg_cost)
-                
-        else:
-            raise ValueError('ERROR : get_cost_for_time_range')
-
-        return timeseries(starttime_sec, req_timestep_sec, data)
-    
-    def get_forecasted_cost_at_time_sec(self, time_sec : float) -> float:
-        '''
-        Description:
-            Given a time in seconds, looksup the data and returns the actual cost of energy at that time. 
-        '''
-        
-        if (time_sec < 0):
-            print("time_sec < 0  should be false: ", time_sec < 0)
-            raise ValueError('ERROR : parameters to get_forecasted_cost_at_time_sec are incompatible')
-
-        time_sec %= self.cost_profile_length_sec
-    
-        return self.forecasted_cost_profile.get_val_from_time(time_sec)
-
-    def get_actual_cost_at_time_sec(self, time_sec : float) -> float:
-        '''
-        Description:
-            Given a time in seconds, looksup the data and returns the actual cost of energy at that time. 
-        '''
-        
-        if (time_sec < 0):
-            print("time_sec < 0  should be false: ", time_sec < 0)
-            raise ValueError('ERROR : parameters to get_actual_cost_at_time_sec are incompatible')
-
-        time_sec %= self.cost_profile_length_sec
-    
-        return self.actual_cost_profile.get_val_from_time(time_sec)
-
-class TE_cost_forecaster_v3():
-    
-    def __init__(
-            self, input_folder: str, figures_folder: str, plot: bool) -> None:
-        
-        loader = load_demand_gen_files(input_folder)                            # loader object
-        (self.dem_dict, self.gen_dict, self.cost_dict) = loader.load()          # loads all input file
-        
-        self.solver = cost_solver(self.cost_dict)
-        
-        self.forecast_dur_s = 48*3600                                           # How long in the future can we forecast
-        self.forecast_ts_s = 1*3600                                             # The timestep in which forecast data is available 
-        self.actual_ts_s = 1*3600                                               # The timestep in which actual data is available
-        self.actual_known_for_time = 0.25*3600
-        
-    def check_time_values(
-            self, start_time_sec: float, end_time_sec: float, 
-            req_time_step_sec: float) -> None:
-        
-        # Ensure time_range doesn't go beyond forecast_duration
-        assert end_time_sec - start_time_sec < self.forecast_dur_s, \
-            "requested time range ({}, {}) hrs is beyond forecast duration of {} hrs"\
-            .format(start_time_sec/3600, end_time_sec/3600, self.forecast_dur_s/3600)
-        
-        # Ensure self.forecast_time_step_sec is a multiple of req_time_step_sec
-        assert abs(fmod(self.actual_ts_s, req_time_step_sec)) < 0.001, \
-            "requested time range ({}, {}) hrs is beyond forecast duration of {} hrs"\
-            .format(start_time_sec/3600, end_time_sec/3600, self.forecast_dur_s/3600)
-        
-        # Ensure start_time_sec is a perfect multiple of req_time_step_sec
-        assert abs(fmod(start_time_sec, req_time_step_sec)) < 0.001 , \
-            "start_time_sec: {} should be a multiple of actual_ts_s: {}"\
-            .format(start_time_sec, self.actual_ts_s)
-        
-        # Ensure end_time_sec is a perfect multiple of time_range
-        assert abs(fmod(end_time_sec, req_time_step_sec)) < 0.001 , \
-            "end_time_sec: {} should be a multiple of req_time_step_sec: {}"\
-            .format(end_time_sec, req_time_step_sec)
-        
-        # Ensure time_step is a perfect multiple of time_range
-        assert abs(fmod(end_time_sec - start_time_sec, req_time_step_sec)) < 0.001 , \
-            "requested time_range_sec: {} should be a multiple of req_time_step_sec: {}"\
-            .format(end_time_sec - start_time_sec, req_time_step_sec)
-    
-    def get_data_for_time_sec( self, data_id: str, data_type: str, cur_time_sec: float):
-        
-        if data_id in self.dem_dict:
-            (metadata_dict, frcst_metadata_dict, data_dict) = self.dem_dict[data_id]
-        elif data_id in self.gen_dict:
-            (metadata_dict, frcst_metadata_dict, data_dict) = self.gen_dict[data_id]
-        else:
-            assert False, "data_id: {} is not present in demand data nor generation data"
-        
-        if data_type == "actual":
-            column_id = "actual"
-            column_ts = self.actual_ts_s
-            
-        elif data_type == "forecast":
-            
-            column_id = None
-            column_ts = self.forecast_ts_s
-            
-            # switch key and values in forecast_metadata_dict so that we can iterate through release_time
-            rel_t_to_frcst_id_d = {y[0]: x for x, y in frcst_metadata_dict.items()}
-            
-            # Start from 1st forecast until the forecast that's active during cur_time_sec
-            for (release_time_hrs, forecast_id) in rel_t_to_frcst_id_d.items():
-                if (release_time_hrs*3600 <= cur_time_sec):
-                    (frcst_rel_t_s, frcst_st_t_s, frcst_end_t_s, frcst_ts_s, frcst_of_s) = frcst_metadata_dict[forecast_id]
-                    
-                    if cur_time_sec >= frcst_st_t_s and cur_time_sec < frcst_end_t_s:
-                        column_id = forecast_id
-                else:
-                    break
-        else:
-            assert False, "Error, check here!!"
-        
-        assert column_id is not None, "column_id is None, that should not happen"
-        
-
-        df = data_dict[column_id]
-        df_time = cur_time_sec - (cur_time_sec % column_ts)
-
-        start =  df["{}_time".format(column_id)].iloc[0]
-        end = df["{}_time".format(column_id)].iloc[-1]
-        
-        #print("column_id:", column_id)
-        #print("df_time:", df_time/3600.0)
-        #print("start:", start/3600.0)
-        #print("end:", end/3600.0)
-        
-        assert df_time >= start and df_time <= end, \
-            "time not in the dataframe"
-        
-        return [float(df[df["{}_time".format(column_id)] == df_time][column_id])]
-    
-    def get_forecast_data_for_time_range(
-            self, data_id: str, cur_time_sec: float, start_time_sec: float, 
-            end_time_sec: float, req_time_step_sec: float):
-        
-        self.check_time_values(start_time_sec, end_time_sec, req_time_step_sec)
-        
-        if data_id in self.dem_dict:
-            (metadata_dict, frcst_metadata_dict, data_dict) = self.dem_dict[data_id]
-        elif data_id in self.gen_dict:
-            (metadata_dict, frcst_metadata_dict, data_dict) = self.gen_dict[data_id]
-        else:
-            assert False, "data_id: {} is not present in demand data nor generation data"
-        
-        final_data = np.zeros(int((end_time_sec - start_time_sec) / req_time_step_sec))
-        
-        # switch key and values in forecast_metadata_dict so that we can iterate through release_time
-        rel_t_to_frcst_id_d = {y[0]: x for x, y in frcst_metadata_dict.items()}
-        
-        # Start from 1st forecast until the forecast that's active during cur_time_sec
-        for (release_time_hrs, forecast_id) in rel_t_to_frcst_id_d.items():
-            if (release_time_hrs*3600 <= cur_time_sec):
-                # frcst_rel_t_s = Forecast Release Time Sec
-                # frcst_st_t_s = Forecast Start Time Sec
-                # frcst_end_t_s = Forecast End Time Sec
-                # frcst_ts_s = Forecast Time Step Sec
-                # frcst_of = Forecast Offset Sec
-                (frcst_rel_t_s, frcst_st_t_s, frcst_end_t_s, frcst_ts_s, frcst_of_s) = frcst_metadata_dict[forecast_id]
-                
-                # Check for overlap between forecast start end and requested start end
-                overlap_start = max(start_time_sec, frcst_st_t_s)
-                overlap_end = min(end_time_sec, frcst_end_t_s)
-                
-                if (overlap_start < overlap_end):
-                    df = data_dict[forecast_id]
-                    
-                    forecast_arr = []
-                    for time in np.arange(overlap_start, overlap_end, req_time_step_sec):
-                        time_in_df_timestep = time - (time%self.forecast_ts_s)
-                        forecast_arr.append(float(df[df["{}_time".format(forecast_id)] == time_in_df_timestep][forecast_id]))
-                     
-                    forecast_arr = np.array(forecast_arr)
-                    forecast_idx = np.arange(int((overlap_start-start_time_sec)/req_time_step_sec), int((overlap_end-start_time_sec)/req_time_step_sec))
-                    np.put(final_data, forecast_idx, forecast_arr)
-            else:
-                break
-        
-        return final_data
-        
-    def get_data_for_time_range(
-            self, data_id:float, cur_time_sec:float, start_time_sec: float, 
-            end_time_sec: float, req_time_step_sec: float):
-        
-        arr = self.get_forecast_data_for_time_range(data_id, cur_time_sec, start_time_sec, end_time_sec, req_time_step_sec)
-        
-        if cur_time_sec >= start_time_sec and cur_time_sec < end_time_sec:
-            
-            idx = int((cur_time_sec-start_time_sec)/req_time_step_sec)
-            arr[idx] = self.get_data_for_time_sec(data_id, "actual", cur_time_sec)[0]
-        
-        return arr
-        
-    def get_cost_for_time_range(
-            self, cur_time_sec:float, start_time_sec: float, 
-            end_time_sec: float, req_time_step_sec: float):
-        
-        dem_data = pd.DataFrame()
-        gen_data = pd.DataFrame()
-        
-        # get all known data
-        dem_data["demand"] = self.get_data_for_time_range("demand", cur_time_sec, start_time_sec, end_time_sec, req_time_step_sec)
-        gen_data["nuclear"] = self.get_data_for_time_range("nuclear", cur_time_sec, start_time_sec, end_time_sec, req_time_step_sec)
-        gen_data["solar"] = self.get_data_for_time_range("solar", cur_time_sec, start_time_sec, end_time_sec, req_time_step_sec)
-        gen_data["wind"] = self.get_data_for_time_range("wind", cur_time_sec, start_time_sec, end_time_sec, req_time_step_sec)        
-        gen_data["fossil_fuel"] = dem_data["demand"] - gen_data["nuclear"] - gen_data["solar"] - gen_data["wind"]
-        
-        gen_data["fossil_fuel"][ np.where(gen_data["fossil_fuel"] < 0.0 )[0] ] = 0.0
-        
-        total_cost_usd = np.zeros(int((end_time_sec - start_time_sec) / req_time_step_sec))
-        
-        for (data_id, arr) in gen_data.items():
-            cost_function = self.solver.solve(arr, self.cost_dict[data_id])
-            total_cost_usd += cost_function*arr*req_time_step_sec/3600.0
-            
-        total_MWh_per_timestep = gen_data.sum(axis=1) * req_time_step_sec/3600.0
-        cost_usd_per_MWh = total_cost_usd/total_MWh_per_timestep
-        cost_usd_per_kWh = cost_usd_per_MWh/1000.0
-        
-        return timeseries(start_time_sec, req_time_step_sec, cost_usd_per_kWh)
-    
-    def get_cost_at_time_sec(self, cost_type:str, time_sec:float):
-        req_time_step_sec = 0.25 * 3600
-        
-        dem_data = pd.DataFrame()
-        gen_data = pd.DataFrame()
-
-        print(cost_type, time_sec)
-        dem_data["demand"] = self.get_data_for_time_sec("demand", cost_type, time_sec)
-        gen_data["nuclear"] = self.get_data_for_time_sec("nuclear", cost_type, time_sec)
-        gen_data["solar"] = self.get_data_for_time_sec("solar", cost_type, time_sec)
-        gen_data["wind"] = self.get_data_for_time_sec("wind", cost_type, time_sec)      
-            
-        gen_data["fossil_fuel"] = dem_data["demand"] - gen_data["nuclear"] - gen_data["solar"] - gen_data["wind"]
-        
-        gen_data["fossil_fuel"][ np.where(gen_data["fossil_fuel"] < 0.0 )[0] ] = 0.0
-        
-        total_cost_usd = 0.0
-        
-        for (data_id, arr) in gen_data.items():
-            cost_function = self.solver.solve(arr, self.cost_dict[data_id])
-            total_cost_usd += cost_function*arr*req_time_step_sec/3600.0
-            
-        total_MWh_per_timestep = gen_data.sum(axis=1) * req_time_step_sec/3600.0
-        cost_usd_per_MWh = total_cost_usd/total_MWh_per_timestep
-        cost_usd_per_kWh = cost_usd_per_MWh/1000.0
-        
-        return cost_usd_per_kWh
-        
-    def get_forecasted_cost_at_time_sec(self, time_sec:float):
-        
-        return self.get_cost_at_time_sec("forecast", time_sec)[0]
-    
-    def get_actual_cost_at_time_sec(self, time_sec:float):
-        
-        return self.get_cost_at_time_sec("actual", time_sec)[0]
-
-class load_demand_gen_files():
-    
-    def __init__(self, input_folder : str) -> None:
-        
-        self.input_folder = input_folder
-        self.file_extension = ".csv"
-        self.demand_files = ["demand"]
-        self.generation_files = ["nuclear", "solar", "wind", "fossil_fuel"]
-        self.current_file_being_read = ""
-        
-    def load(self):
-        
-        demand_dict = {}
-        generation_dict = {}
-        
-        for file in self.demand_files:                                          # load demand files
-             demand_dict[file] = self.load_file(
-                 os.path.join(self.input_folder, file + self.file_extension))
-           
-        for file in self.generation_files:                                      # load generation files
-             generation_dict[file] = self.load_file(
-                 os.path.join(self.input_folder, file + self.file_extension))
-        
-        #self.perform_multifile_error_checks()
-        
-        cost_dict = {}
-        for (gen_type, value) in  generation_dict.items():
-            metadata_dict = generation_dict[gen_type][0]
-            local_cost_dict = {}
-            local_cost_dict["gen_min"] = metadata_dict["gen_min"]
-            local_cost_dict["cost_min"] = metadata_dict["cost_min"]
-            local_cost_dict["gen_max"] = metadata_dict["gen_max"]
-            local_cost_dict["cost_max"] = metadata_dict["cost_max"]
-            local_cost_dict["cost_function"] = metadata_dict["cost_function"]
-            cost_dict[gen_type] = local_cost_dict
-        
-        return (demand_dict, generation_dict, cost_dict)
-    
-    # This is incomplete
-    def perform_multifile_error_checks(self, demand_dict, generation_dict):
-        
-        # Get these info
-        #1. actual_ts_s
-        #2. forecast_ts_s
-        #3. forcast_of_s
-        #4. forecast_l_s
-        
-        all_files = self.demand_files + self.generation_files
-        for (i, (file, value))in eval(dict(demand_dict, **generation_dict).items()):
-            (metadata_dict, forecast_metadata_dict, data_dict) = value
-            
-            if i == 0:    
-                for (forecast_id, data_tup) in forecast_metadata_dict.items():
-                    (release_time_sec, start_time_sec, end_time_sec, time_step_sec, offset_sec) = data_tup
-                    forecast_ts_s = time_step_sec
-                    forcast_of_s = start_time_sec - release_time_sec
-                    forecast_l_s = (end_time_sec - start_time_sec)/time_step_sec
-        pass
-    
-    def load_file(self, file_name : str):
-        
-        self.current_file_being_read = file_name                                # Update file being read
-        
-        df_full = pd.read_csv(file_name)                                        # Read the full csv file
-        df_full = self.perform_df_format_checks(df_full)                        # Ensure file formatting is right
-            
-        metadata_dict = self.extract_metadata_dict(df_full)                     # Extract metadata dictionary
-        forecast_metadata_dict = self.extract_forecast_metadata_dict(df_full)   # Extract forecast metadata dictionary
-        data_dict = self.extract_data_dict(df_full, forecast_metadata_dict)     # Extract data dictionary
-        
-        return_val = self.perform_error_checks(                                 # Perform error checks and return updated data
-            metadata_dict, forecast_metadata_dict, data_dict)
-        
-        return return_val
-    
-    def perform_df_format_checks(self, df_full):
-        
-        self.check_column_headers(df_full)                                      # Check column headers
-        df_full.columns = self.extract_header_without_units(df_full.columns)    # Remove units from column headers
-        
-        return df_full
-        
-    def perform_error_checks(
-            self, metadata_dict, forecast_metadata_dict, data_dict):
-        
-        # Ensure metadata contains gen_min, cost_min, gen_max, cost_max
-
-        required_keys = set()
-        required_keys.add(("cost_min", "$/MWh"))
-        required_keys.add(("cost_max", "$/MWh"))
-        required_keys.add(("gen_min", "MW"))
-        required_keys.add(("gen_max", "MW"))
-        required_keys.add(("cost_function", "str"))
-        
-        updated_metadata_dict = dict()
-        metadata_keys = set()
-        for (key, value) in metadata_dict.items():
-            
-            key_split = key.split("|")
-            assert len(key_split) == 2, \
-                "{}: in metadata column, {} does not contain key and value seperated by |"\
-                    .format(self.current_file_being_read, key)
-                    
-            key_id = key_split[0].strip()
-            key_unit = key_split[1].strip()
-            
-            metadata_keys.add((key_id, key_unit))
-            
-            if key_unit == "$/MWh" : value = float(value)
-            elif key_unit == "MW" : value = float(value)
-            elif key_unit == "str" : value = str(value)
-            else : value = str(value)
-            
-            updated_metadata_dict[key_id] = value
-
-        for req_key in required_keys:
-            
-            assert req_key in metadata_keys, \
-                "{}: metadata does not contain key: {} | {}"\
-                    .format(self.current_file_being_read, req_key[0], req_key[1])
-        
-        for (data_id, timeseries_df) in data_dict.items():
-            
-            if len(timeseries_df) > 0:
-                # Check data is within the bounds of Gen min and Gen max 
-                assert timeseries_df[data_id].min() >= updated_metadata_dict["gen_min"], \
-                    "{}: in {} column, the min data:{} is less than gen_min: {}" \
-                        .format(self.current_file_being_read, data_id, timeseries_df[data_id].min(), updated_metadata_dict["gen_min"])
-                        
-                # Check data is within the bounds of Gen min and Gen max 
-                assert timeseries_df[data_id].max() <= updated_metadata_dict["gen_max"], \
-                    "{}: in {} column, the max data:{} is greater than gen_max: {}" \
-                        .format(self.current_file_being_read, data_id, timeseries_df[data_id].max(), updated_metadata_dict["gen_max"])
-            
-                # Check df doesn't have any missing data
-                assert not timeseries_df.isnull().values.any(), \
-                    "{}: in {} column, there is missing data" \
-                        .format(self.current_file_being_read, data_id)
-            
-        return (updated_metadata_dict, forecast_metadata_dict, data_dict)
-    
-    def extract_data_dict(self, df_full, forecast_metadata_dict) -> dict:
-    
-        data_dict = {}
-        
-        all_keys = set()
-        all_keys.add("actual")
-        all_keys.update(forecast_metadata_dict.keys())
-    
-        for key in all_keys:
-        
-            columns_to_read = [key + "_time", key]
-            
-            data_df = df_full[columns_to_read]
-            data_df = data_df.dropna(how = 'all')
-            data_df[key + "_time"] = data_df[key + "_time"] * 3600.0
-            data_dict[key] = data_df
-    
-        return data_dict
-
-    def extract_header_without_units(self, columns):
-        return [column.split('|')[0].strip() for column in columns]
-        
-    def extract_forecast_metadata_dict(self, df_full):
-        
-        forecast_metadata_dict = {}     # return_val
-       
-        forecast_metadata_columns = ["forecast_id", "forecast_release_time"]
-        df_forecast_metadata = df_full[forecast_metadata_columns]
-        df_forecast_metadata = df_forecast_metadata.dropna(how = 'all')
-        
-        forecast_id_to_release_time_dict = dict(
-            zip(df_forecast_metadata.forecast_id, df_forecast_metadata.forecast_release_time))
-        
-        for forecast_id in forecast_id_to_release_time_dict.keys():
-            
-            release_time_sec = forecast_id_to_release_time_dict[forecast_id]
-            
-            df_forecast_time = df_full[forecast_id + "_time"].dropna(how = 'all')
-            
-            start_time_sec = df_forecast_time.iloc[0] * 3600
-            time_step_sec = (df_forecast_time.iloc[1] - df_forecast_time.iloc[0] ) * 3600
-            end_time_sec = start_time_sec + len(df_forecast_time)*time_step_sec
-            offset_sec = start_time_sec - release_time_sec
-            
-            forecast_metadata_dict[forecast_id] = (release_time_sec, start_time_sec, end_time_sec, time_step_sec, offset_sec)            
-        
-        return forecast_metadata_dict
-    
-    def extract_metadata_dict(self, df_full) -> dict:
-        metadata_cols=["metadata_key", "metadata_value"]
-        
-        metadata_df = df_full[metadata_cols]
-        metadata_df = metadata_df.dropna(how = 'all')
-        metadata_dict = dict(zip(metadata_df.metadata_key, metadata_df.metadata_value))
-        return metadata_dict
-        
-    
-    def check_column_headers(self, df_full):
-        
-        column_headers = df_full.columns.tolist()
-
-        # Note : Please use atleast python 3.7. dictionary is ordered in python 3.7 and above
-        import sys
-        assert sys.version_info >= (3, 7)
-
-        column_name_unit_dict = {}
-        column_name_unit_dict["metadata_key"] = "str"
-        column_name_unit_dict["metadata_value"] = "any"
-        column_name_unit_dict["forecast_id"] = "str"
-        column_name_unit_dict["forecast_release_time"] = "hrs"
-        column_name_unit_dict["actual_time"] = "hrs"
-        column_name_unit_dict["actual"] = "MW"
-        column_name_unit_dict["forecast_time"] = "hrs"
-        column_name_unit_dict["forecast_val"] = "MW"
-
-        # Check for first 6 columns that's defined        
-        for i in range(6):
-    
-            column_name = column_headers[i].split("|")[0].strip()
-            column_unit = column_headers[i].split("|")[1].strip()
-    
-            assert column_name == list(column_name_unit_dict.keys())[i], \
-                "{}: column {} name should be {}"\
-                    .format(self.current_file_being_read, i, list(column_name_unit_dict.keys())[i])
-            assert column_unit == column_name_unit_dict[column_name], \
-                "{}: columns {} unit should be {}"\
-                    .format(self.current_file_being_read, i, column_name_unit_dict[column_name])
-
-        forecast_id_columns_df = df_full["forecast_id | str"]
-
-        # Check remaining column names and units in the file
-        for i in range(6, len(column_headers)):
-    
-            forecast_id_idx = int(i/2) - 3
-            forecast_id = forecast_id_columns_df.iloc[forecast_id_idx]
-        
-            column_name = column_headers[i].split("|")[0].strip()
-            column_unit = column_headers[i].split("|")[1].strip()
-    
-            if i%2 == 0:    # Time column
-                assert column_name == forecast_id + "_time", \
-                    "{}: column {} name should be {}"\
-                        .format(self.current_file_being_read, i, forecast_id + "_time")
-                assert column_unit == "hrs", \
-                    "{}: column {} unit should be {}"\
-                        .format(self.current_file_being_read, i, "hrs")
-    
-            else:           # Value column
-                assert column_name == forecast_id, \
-                    "{}: column {} name should be {}"\
-                        .format(self.current_file_being_read, i, forecast_id)
-                assert column_unit == "MW", \
-                    "{}: column {} unit should be {}"\
-                        .format(self.current_file_being_read, i, "hrs")
-
-class cost_solver():
-    
-    def __init__(self, cost_dict):
-        pass        
-        
-    
-    def solve(self, arr, cost_dict):
-        
-        gen_min = cost_dict["gen_min"]
-        gen_max = cost_dict["gen_max"]
-        cost_min = cost_dict["cost_min"]
-        cost_max = cost_dict["cost_max"]
-        cost_function = cost_dict["cost_function"]
-        
-        if abs(cost_max - cost_min) < 0.001:
-            return_val = self.fixed_cost_solve(arr, cost_min)
-            
-        elif cost_function == "linear":
-            return_val = self.linear_solve(arr, gen_min, gen_max, cost_min, cost_max)
-            
-        elif cost_function == "steep_cubic":
-            return_val = self.steep_cubic_solve(arr, gen_min, gen_max, cost_min, cost_max)
-        
-        elif cost_function == "inverse_s":
-            return_val = self.inverse_s_solve(arr, gen_min, gen_max, cost_min, cost_max)
-        
-        else:
-            assert False, "cost function should be linear, steep_cubic or inverse_s"
-        
-        return return_val
-    
-    def fixed_cost_solve(self, arr, cost):
-        return np.full_like(arr, cost, dtype=float)
-
-    def linear_solve(self, arr, gen_min, gen_max, cost_min, cost_max):
-        return (arr - gen_min)/(gen_max-gen_min)*(cost_max-cost_min)+ cost_min
-    
-    def steep_cubic_solve(self, arr, gen_min, gen_max, cost_min, cost_max):
-        dxdy_at_gen_min = 0
-        dxdy_at_gen_max = 0.35
-        
-        
-        A = [[gen_min**0, gen_min**1,   gen_min**2,     gen_min**3], 
-             [0,          1*gen_min**0, 2*gen_min**1,   3*gen_min**2], 
-             [gen_max**0, gen_max**1,   gen_max**2,     gen_max**3],
-             [0,          1*gen_max**0, 2*gen_max**1,   3*gen_max**2]]
-        
-        b = [[cost_min],
-             [dxdy_at_gen_min],
-             [cost_max],
-             [dxdy_at_gen_max]]
-        
-        C = solve(A, b)
-        
-        return C[0][0] * arr**0 + C[1][0] * arr**1 + C[2][0] * arr**2 + C[3][0] * arr**3
-    
-    def inverse_s_solve(self, arr, gen_min, gen_max, cost_min, cost_max):
-        dxdy_at_gen_min = 0.3
-        dxdy_at_gen_max = 0.3
-        
-        
-        A = [[gen_min**0, gen_min**1,   gen_min**2,     gen_min**3], 
-             [0,          1*gen_min**0, 2*gen_min**1,   3*gen_min**2], 
-             [gen_max**0, gen_max**1,   gen_max**2,     gen_max**3],
-             [0,          1*gen_max**0, 2*gen_max**1,   3*gen_max**2]]
-        
-        b = [[cost_min],
-             [dxdy_at_gen_min],
-             [cost_max],
-             [dxdy_at_gen_max]]
-         
-        C = solve(A, b)
-        
-        return C[0][0] * arr**0 + C[1][0] * arr**1 + C[2][0] * arr**2 + C[3][0] * arr**3
